server/requirements.txt
nats-py==2.9.0
